requirements.txt
python-telegram-bot==20.0a2
starlette==0.19.1
uvicorn[standard]==0.18.2